feat: Split FigletInterface to allow fine grained use of concerns
